+++ SoundPost/SoundPost/Views/HomeView/HomeView.swift
@@ -1,15 +1,8 @@
 import SwiftUI
 
-struct HomeView: View {
-    @State var posts: [PostViewModel] = [
-        PostViewModel.createPreview(uploaderName: "사용자1"),
-        PostViewModel.createPreview(uploaderName: "사용자2"),
-        PostViewModel.createPreview(uploaderName: "사용자3"),
-        PostViewModel.createPreview(uploaderName: "사용자4"),
-        PostViewModel.createPreview(uploaderName: "사용자6"),
-        PostViewModel.createPreview(uploaderName: "사용자7"),
-        PostViewModel.createPreview(uploaderName: "사용자8")
-    ]
+struct HomeView2: View {
+    @EnvironmentObject var authViewModel: AuthViewModel
+    @State private var posts: [PostViewModel] = []
     
     // 한 줄에 하나의 아이템만 표시되도록 그리드 설정
     private let columns = [
@@ -20,13 +13,11 @@ struct HomeView: View {
         NavigationStack {
             ScrollView {
                 LazyVGrid(columns: columns) {
-                    ForEach(posts, id: \.postId) { post in
-                        NavigationLink(destination: PostDetailView(post: post)) {
+                    ForEach(posts.sorted(by: { $0.uploadDate > $1.uploadDate }), id: \.postId) { post in
+                        NavigationLink(destination: PostDetailView2(post: post, quickStartViewModel: QuickStartButtonViewModel(authViewModel: authViewModel, uploader: authViewModel.user))) {
                             VStack(spacing: 0) {
                                 PostView(post: post)
-                                
-                                Divider()
-                                    .padding(.leading)
+                                Divider().padding(.leading)
                             }
                             .contentShape(Rectangle()) // 전체 영역을 탭 가능하게 만듦
                         }
@@ -36,20 +27,18 @@ struct HomeView: View {
             }
             .navigationTitle("홈")
             .navigationBarTitleDisplayMode(.inline)
+            .onAppear {
+                fetchAllPosts()
+            }
         }
     }
-}
-
-#Preview {
-    var posts: [PostViewModel] = [
-        PostViewModel.createPreview(uploaderName: "사용자1"),
-        PostViewModel.createPreview(uploaderName: "사용자2"),
-        PostViewModel.createPreview(uploaderName: "사용자3"),
-        PostViewModel.createPreview(uploaderName: "사용자4"),
-        PostViewModel.createPreview(uploaderName: "사용자6"),
-        PostViewModel.createPreview(uploaderName: "사용자7"),
-        PostViewModel.createPreview(uploaderName: "사용자8")
-    ]
-    HomeView(posts: posts)
-        .accentColor(.primary)
+    
+    // 🔹 Firestore에서 모든 포스트 가져오기
+    private func fetchAllPosts() {
+        FirebaseManager.shared.getAllPosts { fetchedPosts in
+            DispatchQueue.main.async {
+                self.posts = fetchedPosts.map { PostViewModel.createPVMwithPost(post: $0, myId: authViewModel.user?.id ?? "") }
+            }
+        }
+    }
 }
